BackEnd/requirements.txt
asgiref
boto3
botocore
certifi
cffi
charset-normalizer
click
colorama
cryptography
Django
django-rest-knox
djangorestframework
dynaconf
hvac
idna
Jinja2
jmespath
knox
loguru
MarkupSafe
pillow
pycparser
python-dateutil
requests
s3transfer
setuptools
shortuuid
six
sqlparse
tzdata
urllib3
validators
win32-setctime